globalexperts/Source/requirements.txt
backcall==0.1.0
beautifulsoup4==4.7.1
certifi==2019.6.16
chardet==3.0.4
decorator==4.4.0
Django==2.2
django-amazon-ses==4.0.0
django-betterforms==1.2
django-ckeditor==5.7.1
django-crispy-forms==1.9.0
django-crontab==0.7.1
django-filter==2.2.0
django-ipware==3.0.2
django-js-asset==1.2.2
django-simple-history==2.12.0
django-tables2==2.0.3
django-user-agents==0.4.0
django-widget-tweaks==1.4.8
django-widgets-improved==1.5.0
djangoajax==2.3.5
idna==2.8
ipdb==0.12
ipython==7.5.0
ipython-genutils==0.2.0
jedi==0.14.0
mysqlclient==1.4.2.post1
olefile==0.46
parso==0.5.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==4.2.1
prompt-toolkit==2.0.9
ptyprocess==0.6.0
pycoingecko==1.4.0
pycrypto==2.6.1
Pygments==2.4.2
pyotp==2.4.0
python-dateutil==2.8.1
pytz==2019.1
requests==2.22.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.4.1
traitlets==4.3.2
ua-parser==0.10.0
user-agents==2.2.0
wcwidth==0.1.7
